+++ Examples/config.example.json
@@ -1,0 +1,3 @@
+{
+    "token": ""
+}
