+++ Examples/index.js
@@ -1,5 +1,6 @@
 const { commandHandlerClient } = require('../dist/commandHandlerClient.js');
 const { Intents } = require('discord.js');
+const config = require('./config.json');
 
 const client = new commandHandlerClient({
     intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES],
@@ -9,24 +10,7 @@ const client = new commandHandlerClient({
 client.loadCommandOnFile(__dirname + '/commands', 'js');
 
 client.on("messageCreate", (msg) => {
-    if (!msg.content.startsWith(client.prefix) || msg.author.bot) return;
-
-    const args = msg.content.slice(client.prefix.length).trim().split(/ +/);
-    const commandName = args.shift().toLowerCase();
-    const command =
-        client.commands.get(commandName) ||
-        client.commands.find(
-            (cmd) => cmd.aliases && cmd.aliases.includes(commandName)
-        );
-
-    if (!command) return;
-
-    try {
-        command.execute(client, msg, args);
-    } catch (error) {
-        console.error(error);
-    }
-    if (!client.commands.has(commandName)) return;
+    client.commandHandler(msg, client);
 });
 
-client.login("your_bot_token");
+client.login(config.token);
