+++ Examples/index.ts
@@ -1,5 +1,6 @@
 import { commandHandlerClient } from "../dist/commandHandlerClient.js";
 import { Intents } from "discord.js";
+const config = require("./config.json");
 
 const client = new commandHandlerClient({
   intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES],
@@ -9,29 +10,7 @@ const client = new commandHandlerClient({
 client.loadCommandOnFile(__dirname + "/commands", "ts");
 
 client.on("messageCreate", (msg) => {
-  if (!msg.content.startsWith(client.prefix) || msg.author.bot) return;
-
-  const args: string[] = msg.content
-    .slice(client.prefix.length)
-    .trim()
-    .split(/ +/);
-  const shift: any = args.shift();
-  const commandName = shift.toLowerCase();
-
-  const command: any =
-    client.commands.get(commandName) ||
-    client.commands.find(
-      (cmd: any) => cmd.aliases && cmd.aliases.includes(commandName)
-    );
-
-  if (!command) return;
-
-  try {
-    command.execute(client, msg, args);
-  } catch (error) {
-    console.error(error);
-  }
-  if (!client.commands.has(commandName)) return;
+  client.commandHandler(msg, client);
 });
 
-client.login("your_bot_token");
+client.login(config.token);
