+++ README.md
@@ -28,29 +28,7 @@ const client = new commandHandlerClient({
 client.loadCommandOnFile(__dirname + "/commands", "ts"); // else client.loadCommandOnFolder(__dirname + "/commands", "ts");
 
 client.on("messageCreate", (msg) => {
-  if (!msg.content.startsWith(client.prefix) || msg.author.bot) return;
-
-  const args: string[] = msg.content
-    .slice(client.prefix.length)
-    .trim()
-    .split(/ +/);
-  const shift: any = args.shift();
-  const commandName = shift.toLowerCase();
-
-  const command: any =
-    client.commands.get(commandName) ||
-    client.commands.find(
-      (cmd: any) => cmd.aliases && cmd.aliases.includes(commandName)
-    );
-
-  if (!command) return;
-
-  try {
-    command.execute(client, msg, args);
-  } catch (error) {
-    console.error(error);
-  }
-  if (!client.commands.has(commandName)) return;
+  client.commandHandler(msg, client);
 });
 
 client.login("your_bot_token");
@@ -96,24 +74,7 @@ const client = new commandHandlerClient({
 client.loadCommandOnFile(__dirname + "/commands", "js"); // else client.loadCommandOnFolder(__dirname + "/commands", "js");
 
 client.on("messageCreate", (msg) => {
-  if (!msg.content.startsWith(client.prefix) || msg.author.bot) return;
-
-  const args = msg.content.slice(client.prefix.length).trim().split(/ +/);
-  const commandName = args.shift().toLowerCase();
-  const command =
-    client.commands.get(commandName) ||
-    client.commands.find(
-      (cmd) => cmd.aliases && cmd.aliases.includes(commandName)
-    );
-
-  if (!command) return;
-
-  try {
-    command.execute(client, msg, args);
-  } catch (error) {
-    console.error(error);
-  }
-  if (!client.commands.has(commandName)) return;
+  client.commandHandler(msg, client);
 });
 
 client.login("your_bot_token");
