+++ src/commandHandlerClient.ts
@@ -1,4 +1,4 @@
-import { Client, Collection } from "discord.js";
+import { Client, Collection, Message } from "discord.js";
 import * as fs from "fs";
 
 class commandHandlerClient extends Client {
@@ -33,6 +33,32 @@ class commandHandlerClient extends Client {
         this.commands.set(command.name, command);
       }
     }
+  }
+
+  async commandHandler(msg: Message, client: any) {
+    if (!msg.content.startsWith(client.prefix) || msg.author.bot) return;
+
+    const args: string[] = msg.content
+      .slice(client.prefix.length)
+      .trim()
+      .split(/ +/);
+    const shift: any = args.shift();
+    const commandName = shift.toLowerCase();
+
+    const command: any =
+      client.commands.get(commandName) ||
+      client.commands.find(
+        (cmd: any) => cmd.aliases && cmd.aliases.includes(commandName)
+      );
+
+    if (!command) return;
+
+    try {
+      command.execute(client, msg, args);
+    } catch (error) {
+      console.error(error);
+    }
+    if (!client.commands.has(commandName)) return;
   }
 }
